+++ imports/ui/components/navbar/NavBar.css
@@ -1,0 +1,156 @@
+nav {
+    background-color: #000;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    transition: 0.6s;
+    padding: 5px 50px;
+  }
+  
+  nav .logo {
+    position: relative;
+    font-weight: 700;
+    color: #fff;
+    text-decoration: none;
+    font-size: 2em;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+    transition: 0.6s;
+  }
+  nav ul {
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  nav .menu-section ul li {
+    position: relative;
+    list-style: none;
+    font-size: 12px;
+  }
+  nav .menu-section ul li .list-item-group {
+    position: relative;
+    margin: 0 15px;
+    padding-left: 15px;
+    text-decoration: none;
+    color: #fff;
+    letter-spacing: 1px;
+    height: 15px;
+  }
+  .list-item-group span {
+    margin-left: 10px;
+  }
+  
+  nav .account-section .user-account {
+    position: fixed;
+    margin-left: -70px;
+    width: 160px;
+    height: 150px;
+    background: #eee;
+    color: #000;
+    border: 1px solid #000;
+    border-radius: 5px;
+  }
+  nav .account-section .user-account-modal {
+    display: block;
+    padding-left: 5px;
+    
+  }
+  nav .account-section .user-account-modal li {
+    list-style: none;
+    font-size: 14px;
+    border-bottom: 1px solid #000;
+    margin-top: 10px;
+    transition: 0.6s;
+  }
+  nav .account-section .user-account-modal li:hover{
+    list-style: none;
+    font-size: 16px;
+    border-bottom: 1px solid rgb(133, 129, 129);
+  }
+  nav .account-section ul li .list-item-group {
+    position: relative;
+    padding-left: 15px;
+    text-decoration: none;
+    color: #000;
+    letter-spacing: 1px;
+    height: 15px;
+  }
+  nav .account-section ul li button {
+    width: 150px;
+    height: 24px;
+    font-weight: 800;
+    border: 1px solid #000;
+    cursor: pointer;
+    background: #fff;
+    transition: 0.6s;
+  }
+  nav .account-section ul li button:hover {
+    background: #000;
+    color: #fff;
+  }
+  nav .hamburger-icon-section {
+    width: 0px;
+    height: 0px;
+    background-color: #000;
+  }
+  nav .hamburger-icon-section button {
+    width: 0px;
+    height: 0px;
+    background-color: #000;
+    font-size: 0px;
+  }
+  
+  nav .hamburger-icon-section  img {
+    width: 0;
+    height: 0;
+  }
+  .footer {
+    background-color: brown;
+    grid-area: footer;
+  }
+
+
+  @media only screen 
+  and (min-device-width: 120px) 
+  and (max-device-width: 480px)
+  and (-webkit-min-device-pixel-ratio: 2) {
+
+  nav {
+    background-color: #000;
+    display: flex;
+    justify-content: space-between;
+    height: 140px;
+    align-items: center;
+  }
+  nav .logo {
+    position: relative;
+    font-weight: 700;
+    color: #fff;
+    text-decoration: none;
+    font-size: 3em;
+    text-transform: uppercase;
+    letter-spacing: 2px;
+  }
+ 
+  nav .hamburger-icon-section {
+    margin-right: 100px;
+    margin-bottom: 75px;
+  }
+
+  nav .hamburger-icon-section  img {
+    width: 70px;
+    height: 70px;
+    background-color: #000;
+    border-radius: 100%;
+  }
+  nav ul {
+    display: none;
+  }
+  nav .menu-section {
+    display: none;
+  }
+  nav .account-section {
+    display: none;
+  }
+}
