+++ imports/ui/pages/about/AboutPage.css
@@ -9,7 +9,6 @@
   .dev-zone-about {
     margin: auto;
     background: #fff;
-    box-shadow: 5px 5px 30px 0 rgb(0, 0, 0,.7);
     padding-top: 15px;
     height: 240px;
     margin-bottom: 50px;
@@ -33,7 +32,6 @@
     align-items: center;
     background-color: #fff;
     color: #000;
-    box-shadow: 7px 7px 50px 0 rgb(0, 0, 0,.7);
     border-radius: 100%;
     font-size: 10px;
   }
@@ -53,7 +51,6 @@
     margin-left: -100px ;
     margin-top: 80px;
     border-radius: 20px;
-    box-shadow: 6px 6px 40px 0 rgb(0, 0, 0,.6);
     padding: 5px;
     font-size: 10px;
     font-weight: 800;
